don't store the copy hash
Can't use this to verify without resolving the linked node anyway:
unless you've figured out which node the link is allowed to be a
copy of, you don't know which content to hash vs ignore; but once
you've resolved the node (if any) that a link refers to, then you
don't need the copy hash. Might as well just resolve links.